requirements.txt
DateTime~=4.7
numpy~=1.23.0
aiogram~=2.25.2
APScheduler~=3.10.4
telegraph~=2.2.0
psycopg2~=2.9.5
motor~=3.1.1
captcha~=0.4
asyncpg~=0.27.0
aiohttp~=3.10.8
pymongo~=4.3.2
python-dotenv-vault~=0.6.4
